@@ -1,7 +1,7 @@
 {
   "name": "devdocs-global",
   "version": "2.0.0",
-  "description": "\ud83c\udf0d AI-powered multilingual documentation with Lingo integration",
+  "description": "🌍 AI-powered multilingual documentation with Lingo integration",
   "type": "module",
   "private": true,
   "scripts": {
@@ -15,7 +15,7 @@
     "init": "node src/cli/index.js init"
   },
   "dependencies": {
-    "next": "^14.0.0",
+    "next": "^14.2.33",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "framer-motion": "^10.16.0",
